Add new PAM module support for supporting FIDO2 login
This does a bunch of complicated trickery that's annoying in order to
allow not stalling while waiting for the token so that tokens that
remain in machines don't break normal remote usage. Therefore if you
want to use the token in most contexts you have to enter a blank
password first.

Due to a bug currently KDE's lockscreen will not show the prompt for the
pin on the token, I am attempting to address this upstream we'll see
what success I find.

Oddly despite identical configuration, it seems like SDDM has something
that allows it to accept either the password *or* the token pin
immediately so it's behavior differs from the lockscreen. I consider
this a bonus and attempt not to think about it significantly.